RDMA: Prevent IO for child process
RDMA MR(memory region) is not forkable, the VMA(virtual memory area)
of a MR gets empty in a child process. Prevent IO for child process to
avoid server crash.

Viktor pointed out that we need to prevent crash if client is using
commands that use fork, close connection at current step.

So check the process is child or not before any RDMA operation.

Signed-off-by: zhenwei pi <pizhenwei@bytedance.com>