Add evalScriptsDictSize function
The expression `dictSize(evalScriptsDict())` is used
a couple times for memory reporting.  `dictSize` is a macro
without NULL checking and used within a `FMTARG` macro in `server.c`.  Overall, this makes it challenging to stub out when
building without the Lua engine.

This change lifts that expression up to its own function `evalScriptsDictSize`.

Signed-off-by: Evan Wies <evan@neomantra.net>