Unprotect rdb channel when bgsave child fails in dual channel replication
If bgsaveerr is error, there is no need to protect the rdb channel.
The impact of this may be that when bgsave fails, we will protect
the rdb channel for 60s. It may occupy the reference of the repl
buf block, making it impossible to recycle it until we free the
client due to COB or free the client after 60s.

Signed-off-by: Binbin <binloveplay1314@qq.com>