Fix bug where node is sending MEET packets in a loop
If we receive a MEET packet from a known node, we disconnect the
outbound link to force a reconnect and sending of a PING packet so that
the other node recognizes the link as belonging to us.

Also deflaked one of the tests. And improved testing code following PR
comment.

Signed-off-by: Pierre Turin <pieturin@amazon.com>