Dont print lua script body when doing check-rdb
In old versions, if we are saving the replication info on disk, we will
save the lua scripts in the RDB see f11a7585a8498689e8fd1afbcab4fdc2ba38c38f.

Currently check-rdb prints all the aux fields, this including lua scripts.
Although we no longer save lua scripts to RDB file, we may use the new
check-rdb tool to check the old RDB files, and the printing of lua scripts
will pollute the output. If it is an RDB file that abuses lua eval scripts,
it will print a lot of content. At the end, if there is a lua script, we
will print an info to show the number of lua scripts.

This may be a potential breaking change, but i dont think someone is using
it to get the lua scripts output, people should not rely on it and we should
not print lua body anyway.

Signed-off-by: Binbin <binloveplay1314@qq.com>