Add `FUNCTION LOAD` and `FCALL` benchmark
This commit adds two new benhchmarks to `valkey-benchmark` tool.

The `FUNCTION LOAD` benchmark evaluates the performance of loading new
Lua scripts. This benchmark is parameterized by the number of functions
that are present in the script. To control the number of functions we
can use the CLI option `--num-functions-in-script`.

The `FCALL` benchmarks evaluates the perfomance of calling a Lua
function of a previously loaded script. This benchmark is parameterized
by the number of keys that are passed in the `FCALL` command. To
control the number of keys we can use the CLI option
`--num-keys-in-fcall`.

Signed-off-by: Ricardo Dias <ricardo.dias@percona.com>