refactor: move all valkey modules related declarations to `module.h`
In this commit we move all structures and functions declarations related
to Valkey modules from `server.h` to the recently added `module.h` file.

This re-organization makes it easier for new contributers to find the
valkey modules related code, as well as reducing the compilation times
when changes are made to the modules code.

Signed-off-by: Ricardo Dias <ricardo.dias@percona.com>