Embed keys in serverObject and fields in hash entry as SDS type 5
Until now, these embedded strings were SDS type 8, becuase they were
copied as-is from EMBSTR encoded string objects, which always use SDS
type 8. This change allows them to be embedded as SDS type 5, saving
two bytes.

The implementation of embeddeding SDS strings in other structures is
refactored. sdswrite() is a new function to create an SDS representation
into a buffer provided by the caller, instead of allocating it.
sdscopytobuffer() which just copied the layout of an existing sds is
deleted.

The DEBUG SDSLEN command output is changed slightly, to account for
correct allocation sizes for embedded strings and not report the same
memory twice.

Signed-off-by: Viktor Söderqvist <viktor.soderqvist@est.tech>